Allow the use of CMake to compile the example
This add the possibility to use CMake, instead of the Autotools, to
build the example. Autogen.sh is still present and usable, though.

The biggest problem I had with ./autogen.sh was the fact that it takes
between 30 seconds to one minute to get the whole project prepared, on
some ARMv7 boards. CMake takes a few seconds at most, on the same
hardware.

Signed-off-by: Myy <myy@miouyouyou.fr>